% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_equal.R
\name{all_equal}
\alias{all_equal}
\alias{check_all_equal}
\title{Checks if all arguments are equal}
\usage{
all_equal(..., .check = identical)

check_all_equal(..., .check = identical)
}
\arguments{
\item{...}{arguments to be compared}

\item{.check}{alternative function to be used to compare the arguments}
}
\description{
Uses \code{identical()} (or some other comparison function) to check if
every argument is equal to one another.
}
\examples{

df <- tibble::tibble(x = runif(10), y = runif(10))
df2 <- df
df3 <- df
df4 <- df

# To ensure the three variables have been copied
df$z <- 1
df2$z <- 1
df3$z <- 1
df4$z <- 2

all_equal(df, df2, df3)
all_equal(df, df2, df4)

new_df <- check_all_equal(df, df2, df3)
\dontrun{
check_all_equal(df, df4)
}


x <- 1
y <- 1
z <- 1

all_equal(x, y, z, .check = magrittr::equals)

w <- check_all_equal(x, y, z, .check = magrittr::equals)
\dontrun{
check_all_equal(x, y, z, 2, .check = magrittr::equals)
check_all_equal(x, 2, .check = magrittr::equals)
}

}
