% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_penguins.R
\name{clean_penguins}
\alias{clean_penguins}
\title{Get a clean penguins dataset}
\usage{
clean_penguins()
}
\description{
This function generates a cleaner version of the palmerpenguins
dataset with no missing values and only 5 variables: \code{species},
\code{sex}, \code{island}, \code{bill_length} and \code{bill_depth} (without the
units as in the original dataset)
}
