% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dot_arg.R
\name{dot_arg}
\alias{dot_arg}
\alias{has_dot_arg}
\alias{is_dot_arg}
\title{Check whether a call has dots for arguments}
\usage{
has_dot_arg(call)

is_dot_arg(arg)
}
\arguments{
\item{call}{a call to check for \code{.} arguments}

\item{arg}{an argment within a call to check for either \code{.} arguments or
for subsettings applied directly to a \code{.} argument
(e.g. \code{.$x}, \code{.[1]} or \code{.[[1]]} will trigger a \code{TRUE} result).}
}
\description{
Looks at arguments within a call and checks whether
they consist of a \code{.} argument, or an attempt to subset
a \code{.} argument (e.g. \code{.$x} or \code{.[[1]]}).
}
\examples{

has_dot_arg(quote(length(.)))

has_dot_arg(quote(length(x)))

has_dot_arg(quote(length(.$x)))
has_dot_arg(quote(length(.[[1]])))
}
