% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/escalate_warning.R
\name{escalate_warning}
\alias{escalate_warning}
\title{Escalate a warning message to an Error message}
\usage{
escalate_warning(w)
}
\arguments{
\item{w}{a warning object (which has a message inside it)}
}
\description{
Uses the \code{rlang::abort()} function to escalate
a warning message into an Error to prevent code from
running any further.
}
