% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_bits.R
\name{filter_bits}
\alias{filter_bits}
\alias{drop_zeros}
\alias{positives}
\alias{negatives}
\alias{integers}
\title{Filtering functions}
\usage{
drop_zeros(x)

positives(x)

negatives(x)

integers(x)
}
\arguments{
\item{x}{vector of numerics to be filtered}
}
\description{
A few additional functions to facilitate filtering of vectors
}
\examples{
x <- seq(-3, 3, 0.5)
drop_zeros(x)
positives(x)
negatives(x)
integers(x)
}
