% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-rate.R
\name{get-rate}
\alias{get-rate}
\alias{get_rate}
\title{Calculate Rate of Change}
\usage{
get_rate(x, time, unit = "days", historic = "backward")
}
\arguments{
\item{x}{value to calculate the rate of}

\item{time}{date or time variable to be used in the calculations of rate}

\item{unit}{units of time to be used in output}

\item{historic}{accepts \code{"backward"}, \code{"forward"} or \code{"average"} as possible values.
\code{"backward"} means look backwards for the rate, in this case the first
observation will have a rate of \code{NA}. Similarly, \code{"forward"} means look
forward and so the last observation will have a rate of \code{NA}. \code{"average"}
will take the average of these two measurements and return no \code{NA}s}
}
\description{
Calculates the rate of change of a value, provided a time variable.
\code{time} and \code{units} should be compatible with the \code{\link{lubridate}} package
}
\examples{
x <- c(4,2,1,4,6,7,4,2,2,9)
time <- ymd(c("2000-08-23","2000-04-19","2000-10-12","2000-07-22","2000-12-13",
"2000-06-20","2000-01-02","2000-11-12","2000-03-11","2000-08-04"))
get_rate(x,time)

df <- tibble(x,time) \%>\%
mutate(x.rate =get_rate(x,time,historic="average"))

}
