% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_age.R
\name{get_age}
\alias{get_age}
\title{Calculates ages}
\usage{
get_age(DOB, Date)
}
\arguments{
\item{DOB}{vector of \code{Date} objects to be used as the date of
birth}

\item{Date}{vector of \code{Date} objects to be used as the date at
which the age is requested}
}
\description{
Calculates the age of individuals at a given date,
based on their date of birth. This uses \code{\link{lubridate}} functions
to calculate an age, which means that leap years are also taken into account.
}
\examples{
birthdays <- lubridate::dmy(c("01/01/1990", "29/02/1992", "31/12/2000", "15/06/1995"))
get_age(birthdays, lubridate::today())
}
