% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_function_string.R
\name{get_function_string}
\alias{get_function_string}
\title{Simplify a function to a string}
\usage{
get_function_string(f)
}
\arguments{
\item{f}{function to be coerced to string}
}
\description{
Extracts the details of a function as either it's name, or
(if anonymous), some minimal details about it.
}
\examples{
f <- function(x) x + 1

get_function_string(f)
get_function_string(
  function(x) x + 1
)
get_function_string(
  function(x) mean(1:10)
)
get_function_string(
  function(x) stats::mean(1:10)
)
get_function_string(
  function(x) (x + 1)
)
get_function_string(
  function(x) (3 * x + 1) / 2
)
get_function_string(
  function(x) {
    x + 2
  }
)
get_function_string(
  function(x) {
    y <- x + 2
    x + y
  }
)
get_function_string(
  function(x) {
    y <- x + 2
    x + y
    y^2
  }
)
}
