% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git}
\alias{git}
\title{Git in terminal}
\usage{
git(...)
}
\arguments{
\item{...}{git argument(s)}
}
\description{
Shortcut to using \code{git} in terminal.
}
\examples{
\dontrun{
git("add -A")
git("commit -m 'Commit Message'")
git("push")
}

\dontrun{
git("add -A", "commit -m 'Commit Message'", "push")
}

}
