% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_push.R
\name{git_push}
\alias{git_push}
\alias{git}
\title{Git Commit & Push}
\usage{
git(message)
}
\arguments{
\item{message}{git commit message}
}
\description{
Shortcut to commit and push to a git repo. It passes
\code{git commit -a -m "message"} to the terminal, followed by
\code{git push}. This should only be used after repos have been set up
}
