% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_abort.R
\name{glue_abort}
\alias{glue_abort}
\alias{glue_warn}
\title{Signal an error or warning using glue}
\usage{
glue_abort(msg, ..., sub = T)

glue_warn(msg, ..., sub = T)
}
\arguments{
\item{msg}{Message to be passed as either an error or a warning}

\item{...}{additional arguments to be passed to \code{rlang::abort()} or \code{rlang::warn()}}

\item{sub}{should the \code{substitute()} function be wrapped around the elements
of \code{msg}}
}
\description{
These functions act as wrappers around \code{rlang::abort()} and
\code{rlang::warn()} to allow the \code{glue}-style insertion of variables
names. By default, object names are parsed via \code{substitute()},
rather than evaluated, meaning that the objects do not need to be
coerced to strings. This is useful inside functions to grab the
variable name passed by the user
}
\examples{
f <- function(input_argument) {
  glue_abort("Error is with {input_argument}")
}

x <- "Show My Name, Not My String"
\dontrun{
f(x)
f("Show Me")
}

}
