% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_names.R
\name{has_names}
\alias{has_names}
\title{Does an object have names?}
\usage{
has_names(x)
}
\arguments{
\item{x}{object to be checked}
}
\description{
Checks to see if object x has \code{names}
assciated with it
}
\examples{
has_names(3)
has_names(c(y=3))

iris2 <- rlang::set_names(iris,NULL)
has_names(iris)
has_names(iris2)


}
