% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgs-space.R
\name{installed_pkgs_sizes}
\alias{installed_pkgs_sizes}
\title{Profile drive space consumption of installed R packages}
\usage{
installed_pkgs_sizes()
}
\value{
A \code{tbl_df} object has four columns: Package, Version,
pkg_dir, pkg_dir_size, which are package name, package version,
package installed directory, and the space the package consumes
in byte.
}
\description{
Profile drive space consumption of installed R packages
}
