% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgs-space.R
\name{installed_pkgs_space_treemap}
\alias{installed_pkgs_space_treemap}
\title{Tree map of space consumption of installed R packages}
\usage{
installed_pkgs_space_treemap()
}
\value{
A \code{ggplot} object
}
\description{
Tree map of space consumption of installed R packages
}
