% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgs-space.R
\name{installed_pkgs_total_size}
\alias{installed_pkgs_total_size}
\title{Total space consumption of installed R packages}
\usage{
installed_pkgs_total_size()
}
\value{
A character scalar
}
\description{
Total space consumption of installed R packages
}
