% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_blank.R
\name{is_blank}
\alias{is_blank}
\title{Checks if a string is blank}
\usage{
is_blank(x)
}
\arguments{
\item{x}{character vector to be checked for blank-ness}
}
\description{
Checks if a character string is made up of only spaces
}
\examples{
x <- c("red", "blue green", "  ", " ", "", "yellow")
is_blank(x)
}
