% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justify.R
\name{justify}
\alias{justify}
\alias{justify.numeric}
\alias{justify.character}
\alias{justify.logical}
\alias{justify.factor}
\alias{justify.default}
\title{Sets alignment of object, ready for printing}
\usage{
justify(x, ...)

\method{justify}{numeric}(x, d = 3, ...)

\method{justify}{character}(x, align = "left", na.rm = TRUE, ...)

\method{justify}{logical}(x, form = "short", case = "upper", align = "left", na.rm = TRUE, ...)

\method{justify}{factor}(x, ..., numeric = F)

\method{justify}{default}(x, ...)
}
\arguments{
\item{x}{object to be justified}

\item{...}{further arguments passed to or from other methods}

\item{d}{explicit number of decimal places.}

\item{align}{\code{"left"}, \code{"right"} or \code{"centre"} (or just \code{"l"}, \code{"r"} or \code{"c"})
to indicate which alignment to apply.}

\item{na.rm}{should \code{NA}s (or \code{NaN}s) be kept}

\item{form}{should logical be printed in \code{"long"} format (\code{"TRUE"} and \code{"FALSE"}),
\code{"short"} format (\code{"T"} or \code{"F"}) or
\code{"numeric"} format (\code{"1"} or \code{"0"})}

\item{case}{should output be \code{"upper"} (e.g. \code{"TRUE"}),
\code{"lower"} (e.g. \code{"false"}) or \code{"capitalised"} (e.g. \code{"True"})}

\item{numeric}{for factors, should the underlying numeric reference be used,
or the level (if \code{TRUE}, \code{as.character()} is applied, otherwise
\code{as.numeric()} is applied)}
}
\description{
Applies formatting to an object to align it appropriately.

For numerics, strictly rounds to \code{d} decimal places and
lines up the decimal by adding appropriate whitespace. The use of \code{d}
will be taken as the exact number of decimal places. Many other number formatting
functions take arguments as "suggestions" only (e.g. \code{options("digits")}) or
use multiple ways of defining length of output
(e.g. significant figures and decimal places concurrently)
}
\examples{
rnd <- 10^(runif(100,-2,8))
rnd <- rnd*runif(100,-1,1)
justify(rnd,2)

set.seed(1)
all_cols <- colours()
all_cols <- unique(gsub("[0-9]","",all_cols))
smp_cols <- sample(all_cols,20)
justify(smp_cols,align="centre")



x <- c(TRUE,FALSE,NA)

justify(x)
justify(x,"long","lower","right")
justify(x,form="long",align="centre")
justify(x,form="numeric")
justify(x,na.rm=TRUE)


}
