% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_object}
\alias{load_object}
\title{Load an object from Rda file}
\usage{
load_object(rda, name)
}
\arguments{
\item{rda}{A character scalar of Rda file path}

\item{name}{A character scalar of object name}
}
\value{
An object
}
\description{
Load an object from Rda file
}
