% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_unload.R
\name{load_unload}
\alias{load_unload}
\alias{load_packages}
\alias{unload_packages}
\title{Load or Unload packages}
\usage{
load_packages(pkgs, ...)

unload_packages(pkgs, ...)
}
\arguments{
\item{pkgs}{character list of package names}

\item{...}{additional arguments to be passed to \code{library} or \code{detach}}
}
\description{
\code{load_packages} checks if a vector of packages are installed, if not, installs them.
Then loads them ready for use. \code{unload_packages} unloads packages, see \code{\link{detach}}
}
\examples{
\dontrun{
load_packages(c("dplyr","tibble"))
unload_packages(c("dplyr","tibble"))
}

}
