% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{logit}
\alias{logit}
\alias{odds}
\alias{unodds}
\alias{unlogit}
\title{The logit function}
\usage{
odds(p)

unodds(a)

logit(p)

unlogit(a)
}
\arguments{
\item{p}{vector of probabilities between 0 and 1 (exclusive)}

\item{a}{a numeric vector}
}
\description{
Calculates the logit (or log-odds) value of a probability
}
\examples{

probs <- runif(100)
a <- logit(probs)

plot(probs,a,xlim=c(0,1))

probs0 <- unlogit(a)

plot(probs,probs0,xlim=c(0,1),ylim=c(0,1))


}
