% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutils-package.R
\docType{package}
\name{mutils-package}
\alias{mutils}
\alias{mutils-package}
\title{mutils: Various utility functions to make data processing smoother}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A resource of various utility functions that can assist data analysts 
    and data processors to make their work easier; most functions are "nifty" 
    rather than "innovative".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://michaelbarrowman.co.uk/mutils}
  \item \url{https://github.com/MyKo101/mutils}
}

}
\author{
\strong{Maintainer}: Michael Barrowman \email{myko101ab@gmail.com} (\href{https://orcid.org/0000-0003-0718-4482}{ORCID})

}
\keyword{internal}
