% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rep_list.R
\name{rep_list}
\alias{rep_list}
\title{Replicate objects into a list}
\usage{
rep_list(x, n)
}
\arguments{
\item{x}{value to be replicated}

\item{n}{number of replications}
}
\description{
Creates a list of length \code{n} containing \code{x}
as every element.
}
