% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_whitespace.R
\name{set_whitespace}
\alias{set_whitespace}
\title{Set whitespace in character}
\usage{
set_whitespace(x, symbol = "&emsp;")
}
\arguments{
\item{x}{vector of character strings to set the whitespace for}

\item{symbol}{character string to be used as the whitespace.
In general, for HTML output (default), use \code{"&emsp;"} and
for LaTeX, use \code{"\\\\\\\\quad"} (the double escape is required)}
}
\description{
When printing whitespace, some applications will
ignore or remove the whitespace, for example it will
simplify \code{"  "} to \code{" "}. In this instance, it will be
useful to replace the double whitespace with a space
preceding by \code{"symbol"} (or the other way round conditionally).
}
\examples{
x <- c("  a", "b  ", "  c  ")
set_whitespace(x)
}
