% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_path}
\alias{split_path}
\title{Split a file path into a vector of folder names}
\usage{
split_path(x)
}
\arguments{
\item{x}{A character scalar of file path.}
}
\value{
A character vector.
}
\description{
Split a file path into a vector of folder names
}
\details{
The function is adopted from \href{https://stackoverflow.com/questions/29214932/split-a-file-path-into-folder-names-vector}{a discussion on StackOverflow}.
}
