% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stars_pval}
\alias{stars_pval}
\title{Generate significance stars from p-values}
\usage{
stars_pval(
  p.value,
  cutpoints = c(0, 0.001, 0.01, 0.05, 1),
  symbols = c("***", "**", "*", "ns"),
  corr = FALSE,
  na = FALSE,
  ...
)
}
\arguments{
\item{p.value}{numeric vector of p-values}

\item{...}{any parameters passed to \code{\link[stats]{symnum}}.}
}
\value{
A character vector containing the same number of elements as
\code{p-value}, with an attribute "legend" providing the conversion pattern.
}
\description{
Generate significance stars (e.g. '***', '**', '*', 'ns') from p-values.
Inspired by \code{gtools::stars.pval()} with extra arguments and flexibility
and different default.
}
\details{
Mapping from p-value ranges to symbols: \describe{ \item{0 - 0.001}{'***'}
\item{0.001 - 0.01}{'**'} \item{0.01 - 0.05}{'*'}
\item{0.05 - 1.0}{'ns'} }
}
\examples{

p.val <- c(0.0004, 0.0015, 0.013, 0.044, 0.067, 0.24)
stars_pval(p.val)
}
\seealso{
\code{\link[stats]{symnum}}
}
\author{
Mark Chen \email{mjchen.gene@gmail.com}
}
\keyword{misc}
