% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{table}
\alias{table}
\title{Cross Tabulation and Table Creation in pipes}
\usage{
table(x, ...)
}
\arguments{
\item{x}{data within which the \code{\link[base]{table}} is
to be evaluated}

\item{...}{arguments to be passed to \code{\link[base]{table}}}
}
\description{
Wraps the base package's \code{\link[base]{table}} function to allow users
to use this within a pipeline.
}
\examples{
tbl <- data.frame(
  x = c("a", "a", "b", "a", "a", "b", "b", "b", "a"),
  y = c("c", "d", "c", "d", "c", "d", "c", "d", "c")
)
table(tbl$x, tbl$y)
table(tbl, x, y)
tbl \%>\% table(x, y)
}
