% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_my_package.R
\name{update_my_package}
\alias{update_my_package}
\alias{reinstall_my_package}
\title{Updates current package}
\usage{
reinstall_my_package(dir = ".")

update_my_package(
  git_message = NULL,
  update_type = "dev",
  ignore_notes = F,
  dir = ".",
  verbose = F
)
}
\arguments{
\item{dir}{directory of the package (or a sub-directory of one).
If NULL, the current working directory will be used.}

\item{git_message}{git commit message to be passed to the \code{git()} function.
If no \code{git_message} is provided, then the update will not be commited.
If \code{git_message} is supplied, then it will be used as the
\verb{git commit} message if used.}

\item{update_type}{what level of update version to use.
Accepts \code{"major"}, \code{"minor"},\code{"patch"}, or \code{"dev"}. Uses \code{"dev"} as default.}

\item{ignore_notes}{should the Notes returned by running \code{devtools::check()} be ignored?
(Errors and Warnings are never ignored and should be fixed before upload)}

\item{verbose}{logical. \verb{Should update_my_package()} be chatty?}
}
\description{
Wrapper functions to quickly re-install the current package
and/or upload it via a \verb{git commit} command (after appropriate checks).
}
\section{Functions}{
\itemize{
\item \code{reinstall_my_package}: performs a few functions to quickly and automatically
update a package. In simple terms, it performs the following:
\itemize{
\item \code{devtools::document()}
\item \code{rmarkdown::render("README.Rmd")}
\item \code{devtools::install()}
\item \code{devtools::load_all()}
}

\item \code{update_my_package}: is much more thorough than
\code{reinstall_my_package()}. It also ensures the package files are
styled according to the tidyverse style guide (essential as per
\code{usethis::use_tidy_style(strict=T)}. It then performs an \verb{R CMD CHECK}
on the package to ensure it works robustly. If the check passes
(i.e. it doesn't throw an Error, Warning or Note), then it will
update the package version (in the \code{DESCRIPTION} file), and then
\code{commit} and \code{push} the package to \code{git}.
}}

