% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_expr.R
\name{eval_expr}
\alias{eval_expr}
\title{Evaluate an expression within context}
\usage{
eval_expr(expr, data = NULL)
}
\arguments{
\item{expr}{expression to evaluate. Can be a quosure, as with
\code{eval_tidy()} or a function to be applied to \code{data}.}

\item{data}{A data frame or named list or vector to be passed to
\code{eval_tidy}, or an argment to be passed to \code{expr}}
}
\description{
Uses \code{eval_tidy()} to evaluate an expression
within the context of \code{data}.
}
\examples{

tbl <- tibble::tibble(x=1:5)
eval_expr(mean(x),tbl)
eval_expr(. \%>\% dplyr::pull(x) \%>\% mean,tbl)


}
