% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/justify.R
\name{justify}
\alias{justify}
\alias{justify.numeric}
\alias{justify.character}
\title{Sets alignment of object, ready for printing}
\usage{
justify(x, ...)

\method{justify}{numeric}(x, d = 3, ...)

\method{justify}{character}(x, width, align = "left", newline = "\\n", ...)
}
\arguments{
\item{x}{object to be justified}

\item{...}{further arguments passed to or from other methods}

\item{d}{explicit number of decimal places.}

\item{width}{number of characters before a \code{newline} is inserted.}

\item{align}{\code{"left"}, \code{"right"} or \code{"centre"} (or just \code{"l"}, \code{"r"} or \code{"c"})
to indicate which alignment to apply.}

\item{newline}{character string to insert as a newline notation.}
}
\description{
Applies formatting to an object to align it appropriately.

For numerics, strictly rounds to \code{d} decimal places and
lines up the decimal by adding appropriate whitespace. The use of \code{d}
will be taken as the exact number of decimal places. Many other number formatting
functions take arguments as "suggestions" only (e.g. \code{options("digits")}) or
use multiple ways of defining length of output
(e.g. significant figures and decimal places concurrently)

For characters, inserts a \code{newline} between words when
the text overflows above the \code{width}. Whitespace is used to
fill in space around text to ensure correct alignment.
Note that in this context, only spaces, \code{" "}, can separate
words. This means that, in general, punctuation is considered to
be part of the word before it (e.g. \code{"test."} is a
five-letter word and \code{"five-letter"} is an eleven-letter word)
}
\examples{
rnd <- 10^(runif(100,-2,8))
rnd <- rnd*runif(100,-1,1)
justify(rnd,2)

set.seed(1)
all_cols <- colours()
all_cols <- unique(gsub("[0-9]","",all_cols))
smp_cols <- sample(all_cols,30)
smp_cols <- paste0(smp_cols,collapse=", ")
just_cols <- justify(smp_cols,width=40,align="centre")

cat(just_cols)

just_cols_html <- justify(smp_cols,width=50,align="r",newline="<br>")
cat(just_cols_html)

}
