% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_unload.R
\name{load_unload}
\alias{load_unload}
\alias{load_packages}
\alias{unload_packages}
\alias{reload_packages}
\title{Load or Unload packages}
\usage{
load_packages(...)

unload_packages(...)

reload_packages(...)
}
\arguments{
\item{...}{list of packages to be passed loaded or unloaded}
}
\description{
\code{load_packages} checks if a vector of packages are installed, if not, installs them.
Then loads them ready for use. \code{unload_packages} unloads packages, see \code{\link{detach}}
}
\examples{
load_packages(magrittr,purrr,stringi,lubridate)
unload_packages(magrittr,purrr)

reload_packages(stringi,lubridate)


}
