% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signed-log.R
\name{signed-log}
\alias{signed-log}
\alias{signed_log}
\alias{unsigned_log}
\title{Calculates the signed log of a value}
\usage{
signed_log(x, base = exp(1))

unsigned_log(x, base = exp(1))
}
\arguments{
\item{x}{a numeric vector.}

\item{base}{a positive number}
}
\description{
The signed logarithm allows a pseudo-logarithmic re-scaling to
be performed on data that is both positive and negativeis. It defined as:

\deqn{ \textrm{signed_log}_b(x) = \textrm{sign}(x)\log_b(|x| + 1)}
}
\examples{
x <- seq(-10,10,0.01)
y <- signed_log(x)

plot(x,y,type="l")

x0 <- unsigned_log(y)

plot(y,x0,type="l")

plot(x,x0,type="l")
}
