% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_my_package.R
\name{update_my_package}
\alias{update_my_package}
\title{Updates current package}
\usage{
update_my_package(git_message = NULL, run_check = F)
}
\arguments{
\item{git_message}{optional. git commit message to be passed to the \code{\link{git}()} function.
If no \code{message} is provided, then the \code{\link{git}()} function will
not be ran}

\item{run_check}{logical. Should the package be checked?}
}
\description{
Performs a few functions to quickly and automatically
update a package. Only :
\itemize{
\item \code{devtools::document()}
\item \code{rmarkdown::render("README.Rmd")}
\item \code{devtools::install()}
\item \code{devtools::load_all()}
\item \code{devtools::check()} (if requested)
\item \code{git(message)} (if requested)
}
}
